UI - working version/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
pyjwt>=2.6.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
